[IMP] Don't preregister the processors [IMP] Don't expose the LinkedDataAPI
All the calls in regard to data fetching/handling now are on the LinkedRenderStore, hiding the underlying layers.

In the continuous effort of making the rendering more dynamic, the registering of prefixes is also moved to the LinkedRenderStore. This causes the `@context` node to be removed from the internal `schema` object, completing another step in the merging of the data schema and the actual data itself (which is required since some websites don't have their data ontology and actual data separated).