[IMP] Removed default type setting from the API
The original system searched for a type value with the request IRI as the subject on each update, but since most requests contain multiple subject properties, a lot of duplicate types were written into the objects. We now provide a `defaultType` member in the LRS, which is used for resolving render classes, and can be used by the layers above.